+++ src/transformers/combineCases.ts
@@ -1,7 +1,7 @@
 import { DataSet, TransformationOutput } from "./types";
-import { setEquality } from "../utils/sets";
+import { setEquality } from "../lib/utils/sets";
 import { DDTransformerState } from "../transformer-components/DataDrivenTransformer";
-import { getContextAndDataSet } from "../utils/codapPhone";
+import { getContextAndDataSet } from "../lib/codapPhone";
 import { readableName } from "../transformer-components/util";
 import { eraseFormulas, allAttrNames, cloneCollection } from "./util";
